Rename db to firebase
- firebase contains db, functions, storage
