+++ src/network/firestore-service.ts
@@ -1,5 +1,5 @@
 import { Furniture } from "@/data/Furniture";
-import { db, functions, storage } from "./db";
+import { db, functions, storage } from "./firebase";
 import collections from "./collections";
 import { deepCopy } from "./converters";
